+++ packages/ovault-evm/contracts/VaultComposerSync.sol
@@ -195,6 +195,7 @@ contract VaultComposerSync is IVaultComposerSync, ReentrancyGuard {
         _sendParam.minAmountLD = 0;
 
         _send(SHARE_OFT, _sendParam, _refundAddress);
+        emit Deposited(_depositor, _sendParam.to, _sendParam.dstEid, _assetAmount, shareAmount);
     }
 
     /**
@@ -247,6 +248,7 @@ contract VaultComposerSync is IVaultComposerSync, ReentrancyGuard {
         _sendParam.minAmountLD = 0;
 
         _send(ASSET_OFT, _sendParam, _refundAddress);
+        emit Redeemed(_redeemer, _sendParam.to, _sendParam.dstEid, _shareAmount, assetAmount);
     }
 
     /**