+++ examples/lzapp-migration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/lzapp-migration-example
 
+## 0.3.2
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.3.1
 
 ### Patch Changes