+++ examples/oft-upgradeable/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @layerzerolabs/oft-upgradeable-example
 
+## 0.3.4
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.3.3
 
 ### Patch Changes